+++ server/src/modules/arcs/arcs.service.ts
@@ -5,7 +5,6 @@ import { Arc } from '../../entities/arc.entity';
 import { Chapter } from '../../entities/chapter.entity';
 import { Gamble } from '../../entities/gamble.entity';
 import { CreateArcDto } from './dto/create-arc.dto';
-import { UpdateArcImageDto } from './dto/update-arc-image.dto';
 import { MediaService } from '../media/media.service';
 import { MediaOwnerType, MediaPurpose } from '../../entities/media.entity';
 
@@ -141,61 +140,6 @@ export class ArcsService {
       data: gambles,
       total: gambles.length,
     };
-  }
-
-  async updateImage(id: number, imageData: UpdateArcImageDto): Promise<Arc> {
-    const arc = await this.repo.findOne({ where: { id } });
-    if (!arc) {
-      throw new NotFoundException(`Arc with id ${id} not found`);
-    }
-
-    const updateData: Partial<Arc> = {};
-    if (imageData.imageFileName !== undefined) {
-      updateData.imageFileName = imageData.imageFileName;
-    }
-    if (imageData.imageDisplayName !== undefined) {
-      updateData.imageDisplayName = imageData.imageDisplayName || null;
-    }
-
-    // Only update if we have data to update
-    if (Object.keys(updateData).length > 0) {
-      await this.repo
-        .createQueryBuilder()
-        .update(Arc)
-        .set(updateData)
-        .where('id = :id', { id })
-        .execute();
-    }
-
-    const updatedArc = await this.repo.findOne({ where: { id } });
-    if (!updatedArc) {
-      throw new NotFoundException(`Arc with id ${id} not found after update`);
-    }
-    return updatedArc;
-  }
-
-  async removeImage(id: number): Promise<Arc> {
-    const arc = await this.repo.findOne({ where: { id } });
-    if (!arc) {
-      throw new NotFoundException(`Arc with id ${id} not found`);
-    }
-
-    // Use query builder to ensure proper update
-    await this.repo
-      .createQueryBuilder()
-      .update(Arc)
-      .set({
-        imageFileName: null,
-        imageDisplayName: null,
-      })
-      .where('id = :id', { id })
-      .execute();
-
-    const updatedArc = await this.repo.findOne({ where: { id } });
-    if (!updatedArc) {
-      throw new NotFoundException(`Arc with id ${id} not found after update`);
-    }
-    return updatedArc;
   }
 
   /**