+++ server/src/modules/users/dto/update-profile.dto.ts
@@ -3,7 +3,6 @@ import { IsOptional, IsInt, Min, IsEnum } from 'class-validator';
 import { ProfilePictureType } from '../../../entities/user.entity';
 
 export class UpdateProfileDto {
-
   @ApiPropertyOptional({
     description: 'ID of favorite quote from the quote database',
     example: 1,
@@ -34,7 +33,8 @@ export class UpdateProfileDto {
   profilePictureType?: ProfilePictureType;
 
   @ApiPropertyOptional({
-    description: 'ID of the selected character media (when profilePictureType is CHARACTER_MEDIA)',
+    description:
+      'ID of the selected character media (when profilePictureType is CHARACTER_MEDIA)',
     example: 1,
     minimum: 1,
   })