+++ docs/CHANGELOG.md
@@ -3,7 +3,6 @@
 - Redis based rate limiting with caching middle ware for responses; add cache headers to specific routes
 - Update factions page/data; include members and possibly calculate faction switches based on events
 - Add highlight of most popular quote, gamble, and character profile picture 
-- Update profile image selection to have character names on the left and be a sort of in-tab window
 - Add image upload for gambles, events, factions, volumes (need to setup Backblaze + CDN); allow character profile picture choice from characters
 - Update landing page volume showcase to handle other volume covers (single and pairs) 
 - Add proper README at the root of the project; add AGPL v3 license to allow people to copy but force them to open source as well
@@ -13,6 +12,7 @@
 - Profile picture can now select from available character display media or keep default Discord profile
 - Updated users page and profile to load the new modern profile header
 - Normalized spoiler wrapper component by replacing it with timeline spoiler component
+- Normalized entity display to be shown on arcs, gambles, gambles, volumes, and factions 
 ### Notes
 - I really like the new profile picture selection. I was inspired by the og duelingnetwork profile pictures that showed the card art of the iconic monsters.